moved garden-related methods to separate class; added more annotations/docstrings; misc cleanup

+++ lib/nier/save_file.py
@@ -11,7 +11,7 @@ from datetime import datetime, timedelta
 from difflib import unified_diff
 from functools import cached_property
 from pathlib import Path
-from typing import Union, Optional
+from typing import Union, Optional, Iterator, Collection
 
 from construct.lib.containers import ListContainer, Container
 
@@ -31,10 +31,13 @@ class Constructed:
         self._data = data
         self._parsed = parsed or self._construct.parse(data)
 
-    def __getattr__(self, attr: str):
-        return _clean(getattr(self._parsed, attr))
+    def __getitem__(self, key: str):
+        return _clean(self._parsed[key])
 
-    __getitem__ = __getattr__
+    __getattr__ = __getitem__
+
+    def __setitem__(self, key: str, value):
+        self._parsed[key] = value
 
     @cached_classproperty
     def _offsets_and_sizes(cls):
@@ -57,12 +60,22 @@ class Constructed:
         for key, (offset, size) in self._offsets_and_sizes.items():
             yield key, self._data[offset: offset + size]
 
-    def diff(self, other: 'Constructed', max_len: Optional[int] = 30, per_line: int = 20, byte_diff: bool = False):
+    def diff(
+        self,
+        other: 'Constructed',
+        *,
+        max_len: Optional[int] = 30,
+        per_line: int = 20,
+        byte_diff: bool = False,
+        keys: Collection[str] = None,
+    ):
         found_difference = False
         for key, own_raw in self.raw_items():
-            own_val = self[key]
+            if keys and key not in keys:
+                continue
             other_raw = other.raw(key)
             if own_raw != other_raw:
+                own_val = self[key]
                 if not found_difference:
                     found_difference = True
                     print(f'--- {self}')
@@ -70,7 +83,7 @@ class Constructed:
 
                 if isinstance(self, GameData) and key == 'slots':
                     for own_slot, other_slot in zip(self.slots, other.slots):
-                        own_slot.diff(other_slot, max_len, per_line, byte_diff)
+                        own_slot.diff(other_slot, max_len=max_len, per_line=per_line, byte_diff=byte_diff, keys=keys)
                 elif not byte_diff and own_val != own_raw and not isinstance(own_val, (float, int, str)):
                     print(colored(f'@@ {key} @@', 6))
                     a = json.dumps(own_val, cls=PseudoJsonEncoder, sort_keys=True, indent=4).splitlines()
@@ -120,6 +133,11 @@ class Constructed:
                 print(colored('\n{}  {}  {}'.format('=' * 30, key, '=' * 30), 14))
                 self.view(key, per_line, hide_empty, **kwargs)
 
+    def _pprint(self, key: str, val):
+        if isinstance(val, dict):
+            val = json.dumps(val, sort_keys=True, indent=4, cls=PseudoJsonEncoder)
+        print(f'{colored(key, 14)}: {val}')
+
     def pprint(self, unknowns: bool = False, **kwargs):
         last_was_view = False
         for key in self._offsets_and_sizes:
@@ -132,9 +150,7 @@ class Constructed:
             else:
                 if last_was_view:
                     print()
-                if isinstance(val, dict):
-                    val = json.dumps(val, sort_keys=True, indent=4, cls=PseudoJsonEncoder)
-                print(f'{colored(key, 14)}: {val}')
+                self._pprint(key, val)
                 last_was_view = False
 
 
@@ -142,20 +158,27 @@ class GameData(Constructed, construct=Gamedata):
     def __init__(self, data: bytes, path: Path = None):
         super().__init__(data)
         self._path = path
-        self._parsed: Gamedata = self._construct.parse(data)
         self.slots = [SaveFile(slot, i) for i, slot in enumerate(self._parsed.slots, 1)]
 
     @classmethod
-    def load(cls, path: Union[str, Path]) -> 'Gamedata':
+    def load(cls, path: Union[str, Path]) -> 'GameData':
         path = Path(path).expanduser()
+        log.debug(f'Loading game data from path={path.as_posix()}')
         with path.open('rb') as f:
             return cls(f.read(), path)
 
     def save(self, path: Union[str, Path] = None, backup: bool = True):
+        """
+        Save changes.
+
+        :param path: Location where save file should be written (defaults to the path from which this save file was read
+          if :meth:`.load` was used or an explicit path was provided)
+        :param backup: Whether a backup copy of the original save file should be saved
+        """
         path = path or self._path
         if not path:
             raise ValueError(f'A path is required to save {self}')
-        if backup:
+        if backup and path.exists():
             bkp_path = unique_path(path.parent, path.name, '.bkp')
             log.info(f'Creating backup: {bkp_path.as_posix()}')
             shutil.copy(path, bkp_path)
@@ -163,29 +186,28 @@ class GameData(Constructed, construct=Gamedata):
         with Path(path).expanduser().open('wb') as f:
             f.write(self._construct.build(self._build()))
 
-    def __repr__(self):
-        return '++'.join(map(repr, self.slots))
+    def __repr__(self) -> str:
+        return '<GameData[\n{}]>'.format(''.join(map('    {!r}\n'.format, self.slots)))
 
     @property
     def ok(self) -> bool:
         return all(f.ok for f in self.slots)
 
     def __getitem__(self, slot_or_key: Union[int, str]):
-        try:
-            return self.slots[slot_or_key] if isinstance(slot_or_key, int) else getattr(self._parsed, slot_or_key)
-        except AttributeError as e:
-            raise KeyError(slot_or_key) from e
+        return self.slots[slot_or_key] if isinstance(slot_or_key, int) else self._parsed[slot_or_key]
 
     def __iter__(self):
         yield from self.slots
 
 
 class SaveFile(Constructed, construct=Savefile):
-    def __init__(self, slot, num: int):
-        super().__init__(slot.data, slot.value)
+    """Represents one save slot."""
+
+    def __init__(self, slot: Container, num: int):
+        super().__init__(slot.data, slot.value)  # data/value are set by RawCopy for the raw bytes and parsed value
         self._num = num
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         name = self.character if self.name.lower() in self.character.lower() else f'{self.name} ({self.character})'
         return (
             f'<SaveFile#{self._num}[{name}, Lv.{self.level} @ {self.location}][{self.play_time}]'
@@ -193,68 +215,107 @@ class SaveFile(Constructed, construct=Savefile):
         )
 
     @property
-    def ok(self):
+    def ok(self) -> bool:
         return self._parsed.corruptness == 200
 
     @cached_property
-    def play_time(self):
+    def play_time(self) -> str:
         hours, seconds = divmod(int(self._parsed.total_play_time), 3600)
         minutes, seconds = divmod(seconds, 60)
         return f'{hours:01d}:{minutes:02d}:{seconds:02d}'
 
     @cached_property
-    def known_words(self):
+    def known_words(self) -> list[str]:
         return [w for w, v in self.words.items() if v]
 
     @cached_property
-    def location(self):
+    def location(self) -> str:
         loc_part = '_'.join(self._parsed.map.split('_')[1:3])
         return MAP_ZONE_MAP.get(loc_part, self._parsed.map)
 
     @cached_property
-    def garden(self):
-        return [[GardenPlot(plot, r, i) for i, plot in enumerate(row)] for r, row in enumerate(self._parsed.garden)]
+    def garden(self) -> 'Garden':
+        return Garden(self)
 
-    def show_garden(self, func=str, prefix: str = ''):
-        columns = [list(map(func, row)) for row in self.garden]
+    def _pprint(self, key: str, val):
+        if key == 'garden':
+            print(f'{colored(key, 14)}:')
+            self.garden.show(prefix='    ')
+        else:
+            super()._pprint(key, val)
+
+
+class Garden:
+    def __init__(self, save_file: SaveFile):
+        self.save_file: SaveFile = save_file
+        self.plots = [
+            [GardenPlot(plot, r, i) for i, plot in enumerate(row)] for r, row in enumerate(save_file._parsed.garden)
+        ]
+
+    def __getitem__(self, row: int) -> list['GardenPlot']:
+        return self.plots[row]
+
+    def __iter__(self) -> Iterator['GardenPlot']:
+        for row in self.plots:
+            yield from row
+
+    def show(self, func=str, prefix: str = ''):
+        columns = [list(map(func, row)) for row in self.plots]
         row_fmt = '{}{{:>{}s}}  {{:>{}s}}  {{:>{}s}}'.format(prefix, *(max(map(len, col)) for col in columns))
         print('\n'.join(row_fmt.format(*row) for row in zip(*columns)))
 
-    def iter_garden_plots(self):
-        for row in self.garden:
-            yield from row
+    def update(self, dt: datetime = None, hours: int = None, fertilizer: Union[str, int] = None, water: int = None):
+        if dt or hours:
+            self.set_plant_times(dt, hours)
+        if fertilizer or fertilizer == 0:
+            self.set_fertilizer(fertilizer)
+        if water:
+            self.set_water(water)
 
     def set_plant_times(self, dt: datetime = None, hours: int = None):
+        """
+        Sets the plant time for all seeds in this garden.  Does not modify plots where a seed has not been planted.
+
+        :param dt: The specific time to set as the time that all seeds were planted.  Mutually exclusive with ``hours``.
+        :param hours: Number of hours earlier than the current time to set as the time that all seeds were planted.
+          Mutually exclusive with ``dt``.
+        """
         if (dt and hours) or (not dt and not hours):
             raise ValueError(f'set_plant_times() requires ONE of dt or hours')
         dt = dt or (datetime.now() - timedelta(hours=hours))
-        for plot in self.iter_garden_plots():
+        for plot in self:
             if plot._parsed.seed != 255:
                 plot._parsed.time = dt
 
-    def set_fertilizer(self, fertilizer):
-        for plot in self.iter_garden_plots():
+    def set_fertilizer(self, fertilizer: Union[str, int]):
+        """
+        Sets the fertilizer used in all garden plots, regardless of whether a seed has been planted in that plot.
+
+        :param fertilizer: One of ['None', 'Speed Fertilizer', 'Flowering Fertilizer', 'Bounty Fertilizer'] or an
+          integer between 0-3, inclusive.
+        """
+        for plot in self:
             plot._parsed.fertilizer = fertilizer
 
     def set_water(self, water: int):
-        kwargs = {'once': water >= 1, 'twice': water >= 2}
-        for plot in self.iter_garden_plots():
+        kwargs = {'first': water >= 1, 'second': water >= 2}
+        for plot in self:
             if plot._parsed.seed != 255:
                 for key, val in kwargs.items():
                     setattr(plot._parsed.water, key, val)
 
 
 class GardenPlot(Constructed, construct=Plot):
-    def __init__(self, plot, row: int, num: int):
-        super().__init__(plot.data, plot.value)
+    def __init__(self, plot: Container, row: int, num: int):
+        super().__init__(plot.data, plot.value)  # data/value are set by RawCopy for the raw bytes and parsed value
         self._row = row
         self._num = num
 
     @cached_property
-    def watered(self):
+    def watered(self) -> str:
         return ''.join('\u25cb' if v else '\u2715' for k, v in self.water.items() if k != '_flagsenum')
 
-    def __str__(self):
+    def __str__(self) -> str:
         planted = self.time.isoformat(' ') if self.time else None
         plant = SEED_RESULT_MAP.get(self.seed, self.seed)
         plant = 'None' if plant == 255 else plant
@@ -262,14 +323,12 @@ class GardenPlot(Constructed, construct=Plot):
         direction = self.direction.hex()
         return f'\u2039{plant} | F:{fertilizer} | W:{self.watered} | {planted} | {direction}\u203a'
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         planted = self.time.isoformat(' ') if self.time else None
         return (
             f'<GardenPlot[{self._row}x{self._num} @ {planted}, {self.seed} + {self.fertilizer} (watered:'
             f' {self.watered})]({self._data[:-8].hex(" ", -4)})>'
         )
-
-    __serializable__ = __repr__
 
 
 def _build(obj):
